backendapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0